+++ code/packages/rust/ruby-parser/src/_grammar.rs
@@ -34,7 +34,6 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"redo_statement"#.to_string() },
                 GrammarElement::RuleReference { name: r#"retry_statement"#.to_string() },
                 GrammarElement::RuleReference { name: r#"yield_statement"#.to_string() },
-                GrammarElement::RuleReference { name: r#"super_statement"#.to_string() },
                 GrammarElement::RuleReference { name: r#"alias_statement"#.to_string() },
                 GrammarElement::RuleReference { name: r#"undef_statement"#.to_string() },
                 GrammarElement::RuleReference { name: r#"multi_assignment"#.to_string() },
@@ -99,6 +98,7 @@ pub fn parser_grammar() -> ParserGrammar {
             name: r#"def_statement"#.to_string(),
             body: GrammarElement::Sequence { elements: vec![
                 GrammarElement::Literal { value: r#"def"#.to_string() },
+                GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"def_receiver"#.to_string() }) },
                 GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
                 GrammarElement::Optional { element: Box::new(GrammarElement::Sequence { elements: vec![
                         GrammarElement::TokenReference { name: r#"LPAREN"#.to_string() },
@@ -115,12 +115,21 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"ensure_clause"#.to_string() }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 115,
+            line_number: 132,
+        },
+        GrammarRule {
+            name: r#"def_receiver"#.to_string(),
+            body: GrammarElement::Sequence { elements: vec![
+                GrammarElement::RuleReference { name: r#"singleton_receiver"#.to_string() },
+                GrammarElement::Literal { value: r#"."#.to_string() },
+            ] },
+            line_number: 138,
         },
         GrammarRule {
             name: r#"endless_def_statement"#.to_string(),
             body: GrammarElement::Sequence { elements: vec![
                 GrammarElement::Literal { value: r#"def"#.to_string() },
+                GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"def_receiver"#.to_string() }) },
                 GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
                 GrammarElement::Optional { element: Box::new(GrammarElement::Sequence { elements: vec![
                         GrammarElement::TokenReference { name: r#"LPAREN"#.to_string() },
@@ -130,7 +139,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"EQUALS"#.to_string() },
                 GrammarElement::RuleReference { name: r#"expression"#.to_string() },
             ] },
-            line_number: 121,
+            line_number: 147,
         },
         GrammarRule {
             name: r#"class_statement"#.to_string(),
@@ -159,7 +168,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     GrammarElement::Literal { value: r#"end"#.to_string() },
                 ] },
             ] },
-            line_number: 142,
+            line_number: 168,
         },
         GrammarRule {
             name: r#"singleton_receiver"#.to_string(),
@@ -167,7 +176,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"self"#.to_string() },
                 GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
             ] },
-            line_number: 144,
+            line_number: 170,
         },
         GrammarRule {
             name: r#"module_statement"#.to_string(),
@@ -180,7 +189,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 145,
+            line_number: 171,
         },
         GrammarRule {
             name: r#"method_with_block"#.to_string(),
@@ -202,7 +211,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::RuleReference { name: r#"block"#.to_string() },
             ] },
-            line_number: 147,
+            line_number: 173,
         },
         GrammarRule {
             name: r#"block"#.to_string(),
@@ -210,7 +219,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"do_block"#.to_string() },
                 GrammarElement::RuleReference { name: r#"brace_block"#.to_string() },
             ] },
-            line_number: 148,
+            line_number: 174,
         },
         GrammarRule {
             name: r#"do_block"#.to_string(),
@@ -223,7 +232,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 149,
+            line_number: 175,
         },
         GrammarRule {
             name: r#"brace_block"#.to_string(),
@@ -233,7 +242,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Repetition { element: Box::new(GrammarElement::RuleReference { name: r#"statement"#.to_string() }) },
                 GrammarElement::TokenReference { name: r#"RBRACE"#.to_string() },
             ] },
-            line_number: 150,
+            line_number: 176,
         },
         GrammarRule {
             name: r#"block_params"#.to_string(),
@@ -254,7 +263,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"|"#.to_string() },
             ] },
-            line_number: 160,
+            line_number: 186,
         },
         GrammarRule {
             name: r#"return_statement"#.to_string(),
@@ -262,7 +271,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"return"#.to_string() },
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"expression"#.to_string() }) },
             ] },
-            line_number: 162,
+            line_number: 188,
         },
         GrammarRule {
             name: r#"break_statement"#.to_string(),
@@ -270,7 +279,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"break"#.to_string() },
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"expression"#.to_string() }) },
             ] },
-            line_number: 163,
+            line_number: 189,
         },
         GrammarRule {
             name: r#"next_statement"#.to_string(),
@@ -278,17 +287,17 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"next"#.to_string() },
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"expression"#.to_string() }) },
             ] },
-            line_number: 164,
+            line_number: 190,
         },
         GrammarRule {
             name: r#"redo_statement"#.to_string(),
             body: GrammarElement::Literal { value: r#"redo"#.to_string() },
-            line_number: 168,
+            line_number: 194,
         },
         GrammarRule {
             name: r#"retry_statement"#.to_string(),
             body: GrammarElement::Literal { value: r#"retry"#.to_string() },
-            line_number: 172,
+            line_number: 198,
         },
         GrammarRule {
             name: r#"alias_statement"#.to_string(),
@@ -297,7 +306,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
                 GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
             ] },
-            line_number: 183,
+            line_number: 209,
         },
         GrammarRule {
             name: r#"undef_statement"#.to_string(),
@@ -305,7 +314,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"undef"#.to_string() },
                 GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
             ] },
-            line_number: 195,
+            line_number: 221,
         },
         GrammarRule {
             name: r#"yield_statement"#.to_string(),
@@ -313,7 +322,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"yield"#.to_string() },
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"yield_args"#.to_string() }) },
             ] },
-            line_number: 217,
+            line_number: 243,
         },
         GrammarRule {
             name: r#"yield_args"#.to_string(),
@@ -337,15 +346,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         ] }) },
                 ] },
             ] },
-            line_number: 218,
-        },
-        GrammarRule {
-            name: r#"super_statement"#.to_string(),
-            body: GrammarElement::Sequence { elements: vec![
-                GrammarElement::Literal { value: r#"super"#.to_string() },
-                GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"super_args"#.to_string() }) },
-            ] },
-            line_number: 237,
+            line_number: 244,
         },
         GrammarRule {
             name: r#"super_args"#.to_string(),
@@ -369,7 +370,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         ] }) },
                 ] },
             ] },
-            line_number: 238,
+            line_number: 271,
         },
         GrammarRule {
             name: r#"params"#.to_string(),
@@ -383,7 +384,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         ] }) },
                 ] },
             ] },
-            line_number: 267,
+            line_number: 300,
         },
         GrammarRule {
             name: r#"param"#.to_string(),
@@ -404,7 +405,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         ] },
                     ] }) },
             ] },
-            line_number: 312,
+            line_number: 345,
         },
         GrammarRule {
             name: r#"if_statement"#.to_string(),
@@ -421,7 +422,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"else_clause"#.to_string() }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 313,
+            line_number: 346,
         },
         GrammarRule {
             name: r#"elsif_clause"#.to_string(),
@@ -435,7 +436,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 314,
+            line_number: 347,
         },
         GrammarRule {
             name: r#"else_clause"#.to_string(),
@@ -446,7 +447,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 315,
+            line_number: 348,
         },
         GrammarRule {
             name: r#"unless_statement"#.to_string(),
@@ -461,7 +462,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"else_clause"#.to_string() }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 316,
+            line_number: 349,
         },
         GrammarRule {
             name: r#"while_statement"#.to_string(),
@@ -474,7 +475,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 317,
+            line_number: 350,
         },
         GrammarRule {
             name: r#"until_statement"#.to_string(),
@@ -487,7 +488,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 318,
+            line_number: 351,
         },
         GrammarRule {
             name: r#"case_statement"#.to_string(),
@@ -501,7 +502,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"else_clause"#.to_string() }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 341,
+            line_number: 374,
         },
         GrammarRule {
             name: r#"when_clause"#.to_string(),
@@ -520,7 +521,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 342,
+            line_number: 375,
         },
         GrammarRule {
             name: r#"in_clause"#.to_string(),
@@ -535,7 +536,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 364,
+            line_number: 397,
         },
         GrammarRule {
             name: r#"pattern"#.to_string(),
@@ -547,7 +548,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"literal_pattern"#.to_string() },
                 GrammarElement::RuleReference { name: r#"binding_pattern"#.to_string() },
             ] },
-            line_number: 365,
+            line_number: 398,
         },
         GrammarRule {
             name: r#"literal_pattern"#.to_string(),
@@ -557,12 +558,12 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"symbol_literal"#.to_string() },
                 GrammarElement::TokenReference { name: r#"KEYWORD"#.to_string() },
             ] },
-            line_number: 366,
+            line_number: 399,
         },
         GrammarRule {
             name: r#"binding_pattern"#.to_string(),
             body: GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
-            line_number: 367,
+            line_number: 400,
         },
         GrammarRule {
             name: r#"array_pattern"#.to_string(),
@@ -583,7 +584,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RBRACKET"#.to_string() },
             ] },
-            line_number: 368,
+            line_number: 401,
         },
         GrammarRule {
             name: r#"hash_pattern"#.to_string(),
@@ -598,7 +599,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RBRACE"#.to_string() },
             ] },
-            line_number: 369,
+            line_number: 402,
         },
         GrammarRule {
             name: r#"hash_pattern_pair"#.to_string(),
@@ -607,7 +608,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"COLON"#.to_string() },
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"pattern"#.to_string() }) },
             ] },
-            line_number: 370,
+            line_number: 403,
         },
         GrammarRule {
             name: r#"splat_pattern"#.to_string(),
@@ -615,7 +616,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"*"#.to_string() },
                 GrammarElement::Optional { element: Box::new(GrammarElement::TokenReference { name: r#"NAME"#.to_string() }) },
             ] },
-            line_number: 377,
+            line_number: 410,
         },
         GrammarRule {
             name: r#"pin_pattern"#.to_string(),
@@ -623,7 +624,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"^"#.to_string() },
                 GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
             ] },
-            line_number: 382,
+            line_number: 415,
         },
         GrammarRule {
             name: r#"class_pattern"#.to_string(),
@@ -639,7 +640,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RPAREN"#.to_string() },
             ] },
-            line_number: 388,
+            line_number: 421,
         },
         GrammarRule {
             name: r#"begin_statement"#.to_string(),
@@ -655,7 +656,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"ensure_clause"#.to_string() }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 409,
+            line_number: 442,
         },
         GrammarRule {
             name: r#"rescue_clause"#.to_string(),
@@ -673,7 +674,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 418,
+            line_number: 451,
         },
         GrammarRule {
             name: r#"exception_list"#.to_string(),
@@ -684,7 +685,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
                     ] }) },
             ] },
-            line_number: 419,
+            line_number: 452,
         },
         GrammarRule {
             name: r#"ensure_clause"#.to_string(),
@@ -695,7 +696,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 420,
+            line_number: 453,
         },
         GrammarRule {
             name: r#"assignment"#.to_string(),
@@ -719,7 +720,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::RuleReference { name: r#"expression"#.to_string() },
             ] },
-            line_number: 440,
+            line_number: 473,
         },
         GrammarRule {
             name: r#"rightward_assignment"#.to_string(),
@@ -728,14 +729,17 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"=>"#.to_string() },
                 GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
             ] },
-            line_number: 459,
+            line_number: 492,
         },
         GrammarRule {
             name: r#"method_call"#.to_string(),
             body: GrammarElement::Sequence { elements: vec![
                 GrammarElement::Group { element: Box::new(GrammarElement::Alternation { choices: vec![
                         GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
-                        GrammarElement::TokenReference { name: r#"KEYWORD"#.to_string() },
+                        GrammarElement::Sequence { elements: vec![
+                            GrammarElement::NegativeLookahead { element: Box::new(GrammarElement::Literal { value: r#"super"#.to_string() }) },
+                            GrammarElement::TokenReference { name: r#"KEYWORD"#.to_string() },
+                        ] },
                     ] }) },
                 GrammarElement::TokenReference { name: r#"LPAREN"#.to_string() },
                 GrammarElement::Optional { element: Box::new(GrammarElement::Sequence { elements: vec![
@@ -748,7 +752,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"RPAREN"#.to_string() },
                 GrammarElement::Repetition { element: Box::new(GrammarElement::RuleReference { name: r#"dot_call"#.to_string() }) },
             ] },
-            line_number: 470,
+            line_number: 509,
         },
         GrammarRule {
             name: r#"dot_call"#.to_string(),
@@ -771,7 +775,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"block"#.to_string() }) },
             ] },
-            line_number: 471,
+            line_number: 510,
         },
         GrammarRule {
             name: r#"scope_resolution"#.to_string(),
@@ -782,7 +786,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::TokenReference { name: r#"KEYWORD"#.to_string() },
                     ] }) },
             ] },
-            line_number: 479,
+            line_number: 518,
         },
         GrammarRule {
             name: r#"call_arg"#.to_string(),
@@ -801,14 +805,17 @@ pub fn parser_grammar() -> ParserGrammar {
                     GrammarElement::RuleReference { name: r#"expression"#.to_string() },
                 ] },
             ] },
-            line_number: 534,
+            line_number: 573,
         },
         GrammarRule {
             name: r#"method_call_no_paren"#.to_string(),
             body: GrammarElement::Sequence { elements: vec![
                 GrammarElement::Group { element: Box::new(GrammarElement::Alternation { choices: vec![
                         GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
-                        GrammarElement::TokenReference { name: r#"KEYWORD"#.to_string() },
+                        GrammarElement::Sequence { elements: vec![
+                            GrammarElement::NegativeLookahead { element: Box::new(GrammarElement::Literal { value: r#"super"#.to_string() }) },
+                            GrammarElement::TokenReference { name: r#"KEYWORD"#.to_string() },
+                        ] },
                     ] }) },
                 GrammarElement::RuleReference { name: r#"expression"#.to_string() },
                 GrammarElement::Repetition { element: Box::new(GrammarElement::Sequence { elements: vec![
@@ -816,17 +823,17 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"expression"#.to_string() },
                     ] }) },
             ] },
-            line_number: 548,
+            line_number: 590,
         },
         GrammarRule {
             name: r#"expression_stmt"#.to_string(),
             body: GrammarElement::RuleReference { name: r#"expression"#.to_string() },
-            line_number: 549,
+            line_number: 591,
         },
         GrammarRule {
             name: r#"expression"#.to_string(),
             body: GrammarElement::RuleReference { name: r#"ternary"#.to_string() },
-            line_number: 656,
+            line_number: 698,
         },
         GrammarRule {
             name: r#"ternary"#.to_string(),
@@ -839,7 +846,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"expression"#.to_string() },
                     ] }) },
             ] },
-            line_number: 657,
+            line_number: 699,
         },
         GrammarRule {
             name: r#"range"#.to_string(),
@@ -862,7 +869,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         ] }) },
                 ] },
             ] },
-            line_number: 658,
+            line_number: 700,
         },
         GrammarRule {
             name: r#"logical_or"#.to_string(),
@@ -876,7 +883,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"logical_and"#.to_string() },
                     ] }) },
             ] },
-            line_number: 659,
+            line_number: 701,
         },
         GrammarRule {
             name: r#"logical_and"#.to_string(),
@@ -890,7 +897,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"logical_not"#.to_string() },
                     ] }) },
             ] },
-            line_number: 660,
+            line_number: 702,
         },
         GrammarRule {
             name: r#"logical_not"#.to_string(),
@@ -901,7 +908,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         ] }) }) },
                 GrammarElement::RuleReference { name: r#"comparison"#.to_string() },
             ] },
-            line_number: 667,
+            line_number: 709,
         },
         GrammarRule {
             name: r#"comparison"#.to_string(),
@@ -919,7 +926,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"sum"#.to_string() },
                     ] }) },
             ] },
-            line_number: 668,
+            line_number: 710,
         },
         GrammarRule {
             name: r#"sum"#.to_string(),
@@ -933,7 +940,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"term"#.to_string() },
                     ] }) },
             ] },
-            line_number: 669,
+            line_number: 711,
         },
         GrammarRule {
             name: r#"term"#.to_string(),
@@ -947,7 +954,15 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"factor"#.to_string() },
                     ] }) },
             ] },
-            line_number: 670,
+            line_number: 712,
+        },
+        GrammarRule {
+            name: r#"super_expr"#.to_string(),
+            body: GrammarElement::Sequence { elements: vec![
+                GrammarElement::Literal { value: r#"super"#.to_string() },
+                GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"super_args"#.to_string() }) },
+            ] },
+            line_number: 781,
         },
         GrammarRule {
             name: r#"factor"#.to_string(),
@@ -955,6 +970,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Group { element: Box::new(GrammarElement::Alternation { choices: vec![
                         GrammarElement::RuleReference { name: r#"defined_expression"#.to_string() },
                         GrammarElement::RuleReference { name: r#"lambda_literal"#.to_string() },
+                        GrammarElement::RuleReference { name: r#"super_expr"#.to_string() },
                         GrammarElement::RuleReference { name: r#"method_call"#.to_string() },
                         GrammarElement::TokenReference { name: r#"NUMBER"#.to_string() },
                         GrammarElement::TokenReference { name: r#"STRING"#.to_string() },
@@ -986,7 +1002,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"scope_resolution"#.to_string() },
                     ] }) },
             ] },
-            line_number: 706,
+            line_number: 782,
         },
         GrammarRule {
             name: r#"lambda_literal"#.to_string(),
@@ -999,7 +1015,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::RuleReference { name: r#"block"#.to_string() },
             ] },
-            line_number: 725,
+            line_number: 801,
         },
         GrammarRule {
             name: r#"unary_minus"#.to_string(),
@@ -1007,7 +1023,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"MINUS"#.to_string() },
                 GrammarElement::RuleReference { name: r#"factor"#.to_string() },
             ] },
-            line_number: 726,
+            line_number: 802,
         },
         GrammarRule {
             name: r#"defined_expression"#.to_string(),
@@ -1015,7 +1031,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"defined?"#.to_string() },
                 GrammarElement::RuleReference { name: r#"factor"#.to_string() },
             ] },
-            line_number: 737,
+            line_number: 813,
         },
         GrammarRule {
             name: r#"symbol_literal"#.to_string(),
@@ -1027,7 +1043,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::TokenReference { name: r#"STRING"#.to_string() },
                     ] }) },
             ] },
-            line_number: 744,
+            line_number: 820,
         },
         GrammarRule {
             name: r#"array_literal"#.to_string(),
@@ -1042,7 +1058,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RBRACKET"#.to_string() },
             ] },
-            line_number: 745,
+            line_number: 821,
         },
         GrammarRule {
             name: r#"hash_literal"#.to_string(),
@@ -1057,7 +1073,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RBRACE"#.to_string() },
             ] },
-            line_number: 746,
+            line_number: 822,
         },
         GrammarRule {
             name: r#"hash_entry"#.to_string(),
@@ -1077,7 +1093,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     GrammarElement::RuleReference { name: r#"expression"#.to_string() },
                 ] },
             ] },
-            line_number: 747,
+            line_number: 823,
         },
     ],
         version: 1,